Backend/requirements.txt
Flask==2.2.2
pymongo==4.3.3
PyJWT==2.6.0
werkzeug==2.2.2
python-dotenv==0.21.0
bcrypt==4.0.1
pytest==7.4.3
pytest-cov==4.1.0
